Add demo dotnet 7, rename old demo, 32 and 64 version
