--- DotNet/Demos/Demo_ConsoleApp_F461/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Tracetool.DotNet.Api" version="13.0.1" targetFramework="net461" />
-</packages>
