+++ DotNet/Demos/Demo_wpf/Window1.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="Vs2017_Wpf_Demo.Window1"
+﻿<Window x:Class="Wpf_Demo.Window1"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     Title="Demo (WPF) for TraceTool Dot Net Standard 2.0" Height="657.34" Width="751.654" VerticalAlignment="Top" HorizontalAlignment="Left" Loaded="Window_Loaded" Closed="Window_Closed" IsEnabled="True" Icon="App.ico">
