+++ DotNet/Demos/Demo_wpf/packages.config
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Tracetool.DotNet.Api" version="13.0.1" targetFramework="net461" />
+  <package id="Tracetool.DotNet.Api" version="13.0.1" targetFramework="net461" requireReinstallation="true" />
 </packages>
